Add files via upload
The Keylogger project is a Python-based application designed to monitor and record keystrokes on a computer. It captures keyboard inputs in real-time and logs them into a file for review. This project demonstrates how to use Python for monitoring user activities, employing modules like pynput to intercept keyboard events. The keylogger can be customized for various purposes, such as user activity monitoring or conducting cybersecurity exercises to detect unauthorized access. Proper ethical guidelines and permissions must be observed when using or deploying this tool.